+++ Carga_hiscred_v1.py
@@ -20,16 +20,24 @@ warnings.simplefilter('error', UserWarning)
 #funciones
 def logging_carga(cursor_con, filename, staging_table):
     mysql_log_load = "insert into Operacion_datamart.Importacion "
-    mysql_log_load += " select curdate() as fecha, '" + filename + "' as nombre_archivo, user() as Usuario, count(*) as Registros from Staging." + staging_table + "; ";
-    cursor_con.execute(mysql_log_load)
+    mysql_log_load += " select curdate() as fecha, '" + filename + "' as nombre_archivo, user() as Usuario, count(*) as Registros, 0 as reproceso from Staging." + staging_table + "; ";
+    cursor.execute(mysql_log_load)
     
 def logging_proceso(cursor_con, process, total_steps, step, descripcion):
     Etapa = str(step) + "/" + str(total_steps) + ") " + descripcion
     mysql_log_task = "insert into Operacion_datamart.Logs_procesos (fecha, Proceso, Etapa) "
     mysql_log_task += " select now() as fecha, '" + process + "' , '" + Etapa + "';"
     #print(mysql_log_task)
-    cursor_con.execute(mysql_log_task)
+    cursor.execute(mysql_log_task)
     
+def Validacion_archivo (filename):
+    stmt = "select * from Operacion_datamart.Importacion where Nombre_archivo = '" + filename + "' and Reproceso = 0;"
+    cursor.execute(stmt)
+    result = cursor.fetchone()
+    if result:
+        return True
+    else:
+        return False    
 
 #Constantes
 date1 = '2020-04-04'  
@@ -75,6 +83,11 @@ for date in datelist:
                               autocommit=True,
                               local_infile=1)
             cursor = con.cursor()
+            if Validacion_archivo(filename):
+                print('Archivo previamente cargado: ' + filename)
+                con.close()
+                continue
+
             cursor.execute("truncate table Staging." + table + ';')
             cursor.execute(load_sql)
             logging_carga(cursor, filename, table)
